+++ src/Client/IPFSClient.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace IPFS\Client;
 
 use IPFS\Exception\IPFSTransportException;
+use IPFS\Helper\FilesystemHelper;
 use IPFS\Model\File;
+use IPFS\Model\Directory;
 use IPFS\Model\ListFileEntry;
 use IPFS\Model\Node;
 use IPFS\Model\Peer;
@@ -14,12 +16,15 @@ use IPFS\Model\Version;
 use IPFS\Transformer\FileLinkTransformer;
 use IPFS\Transformer\FileListTransformer;
 use IPFS\Transformer\FileTransformer;
+use IPFS\Transformer\DirectoryTransformer;
 use IPFS\Transformer\NodeTransformer;
 use IPFS\Transformer\PeerIdentityTransformer;
 use IPFS\Transformer\PeerStreamTransformer;
 use IPFS\Transformer\PeerTransformer;
 use IPFS\Transformer\PingTransformer;
 use IPFS\Transformer\VersionTransformer;
+use Symfony\Component\Mime\Part\DataPart;
+use Symfony\Component\Mime\Part\Multipart\FormDataPart;
 
 class IPFSClient
 {
@@ -61,6 +66,99 @@ class IPFSClient
 
         $transformer = new FileTransformer();
         return $transformer->transform($parsedResponse);
+    }
+
+    public function addFile(string $path, array $parameters = []): File
+    {
+        if (is_file($path) === false) {
+            throw new \InvalidArgumentException('Path must be a file.');
+        }
+        $parameters['wrap-with-directory'] = false;
+
+        $response = $this->httpClient->request('POST', '/api/v0/add', [
+            'body' => [
+                'file' => fopen($path, 'r'),
+            ],
+            'query' => $parameters,
+            'headers' => [
+                'Content-Type' => 'multipart/form-data',
+            ],
+        ]);
+
+        $parsedResponse = json_decode($response, true);
+
+        $transformer = new FileTransformer();
+        return $transformer->transform($parsedResponse);
+    }
+
+    public function addDirectory(string $path, array $parameters = []): Directory
+    {
+        if (is_dir($path) === false) {
+            throw new \InvalidArgumentException('Path must be a directory.');
+        }
+        $parameters['wrap-with-directory'] = true;
+
+        $files = FilesystemHelper::listFiles($path);
+
+        $dataParts = [];
+        $basePath = realpath($path);
+        if ($basePath === false) {
+            throw new \UnexpectedValueException('Path ' . $path . ' does not exist.');
+        }
+
+        foreach ($files as $filePath) {
+            $filePathReal = realpath($filePath);
+            if ($filePathReal === false) {
+                throw new \UnexpectedValueException('Path ' . $path . ' does not exist.');
+            }
+
+            if (str_starts_with($filePathReal, $basePath) === false) {
+                throw new \RuntimeException("File $filePath is outside of base path.");
+            }
+
+            $relativePath = mb_substr($filePathReal, mb_strlen($basePath) + 1); // +1 to remove leading slash
+            $relativePath = str_replace(\DIRECTORY_SEPARATOR, '/', $relativePath); // For Windows
+
+            $fileHandle = fopen($filePath, 'r');
+            if ($fileHandle === false) {
+                throw new \RuntimeException('Unable to open file ' . $filePath);
+            }
+
+            $dataParts[] = new DataPart($fileHandle, $relativePath);
+        }
+
+        // Use Symfony's FormDataPart to build the body + headers
+        $formData = new FormDataPart([
+            'file' => $dataParts,
+        ]);
+
+        $response = $this->httpClient->request('POST', '/api/v0/add', [
+            'body' => $formData->bodyToString(),
+            'query' => $parameters,
+            'headers' => $formData->getPreparedHeaders()->toArray(),
+        ]);
+
+        $parts = explode("\n", $response);
+        $filtered = array_filter($parts, function (string $value) {
+            return mb_strlen(trim($value)) > 0;
+        });
+        $deserializedParts = array_map(fn (string $part) => json_decode($part, true), $filtered);
+        // Sort by size, larger first
+        usort($deserializedParts, function ($a, $b) {
+            return (int) $a['Size'] < (int) $b['Size'] ? 1 : -1;
+        });
+
+        $directoryTransformer = new DirectoryTransformer();
+
+        $directoryData = array_shift($deserializedParts);
+        $directory = $directoryTransformer->transform($directoryData);
+
+        $fileTransformer = new FileTransformer();
+        foreach ($deserializedParts as $part) {
+            $directory->addFile($fileTransformer->transform($part));
+        }
+
+        return $directory;
     }
 
     public function cat(string $hash, int $offset = 0, ?int $length = null): string